Feature/sdl3 port (#69)
* Ports from SDL2 to SDL3

Sets the default C version to C99

Re-worked controller handling
    The controller handling has been improved to allow for a more
    seamless connect/disconnect experience. Controllers are now loaded
    (if rezognized) when they are connected rather then at game boot.

    The tooltip system has been re-written to allow for loading 3 types
    of tooltips in one sitting. One for each controller type. The cost
    is some extra sprite memory. But in all fairness this is not a
    massiv impact on the game.

    This change introduced the `typedef struct tooltip Tooltip` as well
    as `enum GamepadType`.

+++ src/timer.c
@@ -16,7 +16,7 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-#include <SDL.h>
+#include <SDL3/SDL.h>
 #include <stdlib.h>
 #include "util.h"
 #include "timer.h"
@@ -43,7 +43,7 @@ bool timer_started(Timer *t)
 	return t->startTime != 0;
 }
 
-unsigned int timer_get_ticks(Timer *t)
+Uint64 timer_get_ticks(Timer *t)
 {
 	if (!t->startTime)
 		return 0;
